requirements.txt
numpy==1.23.3
Pillow==9.2.0
psd_tools==1.9.23
psd_tools.egg==info
psutil==5.9.2
pyrsistent==0.18.1
